saleCover/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
waitress==2.1.2
whitenoise==6.2.0
